requirement.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.4.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
loguru==0.7.3
packaging==24.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
